Split delegating decorator into two classes
For readability, `CacheCleanupFileAccessTimeProvidingCleanupAction` is
split into two classes:

 - `SnapshottingCacheCleanupFileAccessTimeProvider` which is passed to
   other `CleanupAction` implementations
 - `DelegatingCleanupAction` which takes care of delegating to the main
   `CleanupAction` and closing the `CacheCleanupFileAccessTimeProvider`
